requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
WTForms
email-validator
bcrypt
pdfplumber
reportlab
python-dotenv
gunicorn
mistune
xhtml2pdf
flask_mail
twilio
requests
duckduckgo-search
beautifulsoup4
readability-lxml
httpx
langdetect
fastapi
whoosh
sentence-transformers
hnswlib
llama-cpp-python
langchain